#2 refactor. 디렉토리 정렬
