+++ src/components/ui/Layout.tsx
@@ -6,10 +6,10 @@ interface LayoutProps {
 }
 const Layout = (props: LayoutProps) => {
   let styles =
-    "bg-cover bg-[url('/img/main_background.png')] bg-clubfest-blue overflow-hidden relative h-full flex flex-col max-w-[500px] w-full justify-center items-center ";
+    "bg-cover bg-[url('/src/assets/img/main_background.png')] bg-clubfest-blue overflow-hidden relative h-full flex flex-col max-w-[500px] w-full justify-center items-center ";
   if (props.bg === "landing") {
     styles =
-      "bg-center bg-no-repeat bg-cover bg-[url('/img/landing_background.png')] overflow-hidden relative  flex flex-col max-w-[500px] w-full h-full justify-center items-center";
+      "bg-center bg-no-repeat bg-cover bg-[url('/src/assets/img/landing_background.png')] overflow-hidden relative  flex flex-col max-w-[500px] w-full h-full justify-center items-center";
   }
   return <div className={styles}>{props.children}</div>;
 };
